requirements.txt
hexbytes==0.3.1
pandas==2.1.2
python-dotenv==1.0.1
requests==2.31.0
web3==6.14.0
